backend/requirements.txt
django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-multipart==0.0.6
requests==2.31.0
openai==0.28.0
python-dotenv==1.0.0
pgvector==0.2.4
PyPDF2==3.0.1
numpy==1.26.3
gunicorn==21.2.0
langchain==0.0.325
tiktoken==0.5.2
